# core/vm: opt opSwap*